move changeset-version file
