+++ .github/workflows/changesets.yml
@@ -46,7 +46,7 @@ jobs:
         id: changesets
         uses: changesets/action@v1
         with:
-          version: node .github/changeset-version.js
+          version: node tools/changeset-version.js
           publish: pnpm exec changeset publish
         env:
           GITHUB_TOKEN: ${{ secrets.GH_ACCESS_TOKEN }}
