+++ .github/workflows/e2e.yml
@@ -44,7 +44,7 @@ jobs:
           turbo-signature: ${{ secrets.TURBO_REMOTE_CACHE_SIGNATURE_KEY }}
 
       - name: Bump package versions
-        run: node .github/changeset-version.js
+        run: node tools/changeset-version.js
         env:
           GITHUB_TOKEN: ${{ github.token }}
 
@@ -107,7 +107,7 @@ jobs:
           turbo-signature: ${{ secrets.TURBO_REMOTE_CACHE_SIGNATURE_KEY }}
 
       - name: Bump package versions
-        run: node .github/changeset-version.js
+        run: node tools/changeset-version.js
         env:
           GITHUB_TOKEN: ${{ github.token }}
