+++ .github/workflows/test-and-check.yml
@@ -39,7 +39,7 @@ jobs:
           turbo-signature: ${{ secrets.TURBO_REMOTE_CACHE_SIGNATURE_KEY }}
 
       - name: Bump package versions
-        run: node .github/changeset-version.js
+        run: node tools/changeset-version.js
         env:
           GITHUB_TOKEN: ${{ github.token }}
 
@@ -135,7 +135,7 @@ jobs:
           node-version: ${{ matrix.node_version }}
 
       - name: Bump package versions
-        run: node .github/changeset-version.js
+        run: node tools/changeset-version.js
         env:
           GITHUB_TOKEN: ${{ github.token }}
