+++ tools/deployments/__tests__/changeset-version.test.ts
@@ -1,6 +1,6 @@
 import { expect, it } from "vitest";
 // @ts-expect-error This is a JS file
-import { getNextMiniflareVersion } from "../../../.github/changeset-version";
+import { getNextMiniflareVersion } from "../../changeset-version";
 
 // prettier-ignore
 const miniflareVersionTestCases = [
